pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "pandas>=2.3.3",
    "python-nmap>=0.7.1",
    "requests>=2.32.5",
]
